requirements.txt
openpyxl>=3.1.0
reportlab>=4.0.0
xlrd>=2.0.0
tkinterdnd2>=0.3.0
python-docx>=1.1.0